Unify all buildApi related functions
