+++ packages/codemirror-blocks/src/ui/TextEditor.tsx
@@ -3,37 +3,19 @@ import {
   IUnControlledCodeMirror,
   UnControlled as CodeMirror,
 } from "react-codemirror2";
-import { API } from "./ToggleEditor";
 import { CodeMirrorFacade } from "../editor";
-
-// CodeMirror APIs that we need to disallow
-// NOTE(Emmanuel): we should probably block 'on' and 'off'...
-const unsupportedAPIs = ["startOperation", "endOperation", "operation"];
-
-const buildAPI = () => {
-  const api = {};
-  // show which APIs are unsupported
-  unsupportedAPIs.forEach(
-    (f) =>
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      ((api as any)[f] = () => {
-        throw `The CM API '${f}' is not supported in CodeMirrorBlocks`;
-      })
-  );
-  return api as API;
-};
 
 type Props = {
   codemirrorOptions?: CodeMirror.EditorConfiguration;
   value: string;
   onBeforeChange?: IUnControlledCodeMirror["onBeforeChange"];
-  onMount: (ed: CodeMirrorFacade, api: API) => void;
+  onMount: (ed: CodeMirrorFacade) => void;
 };
 
 const TextEditor = (props: Props) => {
   // build the API on mount
   const handleEditorDidMount = (ed: CodeMirror.Editor) => {
-    props.onMount(new CodeMirrorFacade(ed), buildAPI());
+    props.onMount(new CodeMirrorFacade(ed));
   };
 
   // Build the API for a text editor, restricting APIs that are
