+++ packages/codemirror-blocks/src/ui/ToggleEditor.tsx
@@ -8,179 +8,18 @@ import { mountAnnouncer, say } from "../announcer";
 import TrashCan from "./TrashCan";
 import type { Language, Options } from "../CodeMirrorBlocks";
 import CodeMirror from "codemirror";
-import type { BuiltAPI as BlockEditorAPIExtensions } from "../CodeMirror-api";
+import { buildAPI, API } from "../CodeMirror-api";
 import { CodeMirrorFacade } from "../editor";
 import { AppContext } from "../components/Context";
 import { useDispatch, useSelector, useStore } from "react-redux";
 import * as selectors from "../state/selectors";
 import * as actions from "../state/actions";
-import { AppDispatch, AppStore } from "../state/store";
+import { AppDispatch } from "../state/store";
 
 const defaultCmOptions: CodeMirror.EditorConfiguration = {
   lineNumbers: true,
   viewportMargin: 10,
   extraKeys: { "Shift-Tab": false },
-};
-
-// This is the complete list of methods exposed by the CodeMirror object
-// SOME of them we override, but many can be exposed directly
-// See buildAPI() in the ToggleEditor component
-const codeMirrorAPI = [
-  "getValue",
-  "setValue",
-  "getRange",
-  "replaceRange",
-  "getLine",
-  "lineCount",
-  "firstLine",
-  "lastLine",
-  "getLineHandle",
-  "getLineNumber",
-  "eachLine",
-  "markClean",
-  "changeGeneration",
-  "isClean",
-  "getSelection",
-  "getSelections",
-  "replaceSelection",
-  "replaceSelections",
-  "getCursor",
-  "listSelections",
-  "somethingSelected",
-  "setCursor",
-  "setSelection",
-  "setSelections",
-  "addSelection",
-  "extendSelection",
-  "extendSelections",
-  "extendSelectionsBy",
-  "setExtending",
-  "getExtending",
-  "hasFocus",
-  "findPosH",
-  "findPosV",
-  "findWordAt",
-  "setOption",
-  "getOption",
-  "addKeyMap",
-  "removeKeyMap",
-  "addOverlay",
-  "removeOverlay",
-  "on",
-  "off",
-  "undo",
-  "redo",
-  "undoSelection",
-  "redoSelection",
-  "historySize",
-  "clearHistory",
-  "getHistory",
-  "setHistory",
-  "markText",
-  "setBookmark",
-  "findMarks",
-  "findMarksAt",
-  "getAllMarks",
-  "setGutterMarker",
-  "clearGutter",
-  "addLineClass",
-  "removeLineClass",
-  "lineInfo",
-  "addWidget",
-  "addLineWidget",
-  "setSize",
-  "scrollTo",
-  "getScrollInfo",
-  "scrollIntoView",
-  "cursorCoords",
-  "charCoords",
-  "coordsChar",
-  "lineAtHeight",
-  "heightAtLine",
-  "defaultTextHeight",
-  "defaultCharWidth",
-  "getViewport",
-  "refresh",
-  "operation",
-  "startOperation",
-  "endOperation",
-  "indentLine",
-  "toggleOverwrite",
-  "isReadOnly",
-  "lineSeparator",
-  "execCommand",
-  "posFromIndex",
-  "indexFromPos",
-  "focus",
-  "phrase",
-  "getInputField",
-  "getWrapperElement",
-  "getScrollerElement",
-  "getGutterElement",
-] as const;
-
-type CodeMirrorAPI = Pick<CodeMirror.Editor, typeof codeMirrorAPI[number]>;
-
-type ToggleEditorAPI = {
-  getBlockMode(): boolean;
-  setBlockMode(blockMode: boolean): void;
-  getCM(): CodeMirror.Editor;
-  on: CodeMirror.Editor["on"];
-  off: CodeMirror.Editor["off"];
-  runMode(): never;
-};
-
-export type API = ToggleEditorAPI & CodeMirrorAPI & BlockEditorAPIExtensions;
-
-/**
- * @internal
- * Populate a base object with mode-agnostic methods we wish to expose
- */
-const buildAPI = (
-  editor: CodeMirrorFacade,
-  store: AppStore,
-  language: Language,
-  eventHandlers: Record<string, ((...args: unknown[]) => void)[]>
-): ToggleEditorAPI & Partial<CodeMirrorAPI> => {
-  const base: Partial<CodeMirrorAPI> = {};
-  // any CodeMirror function that we can call directly should be passed-through.
-  // TextEditor and BlockEditor can add their own, or override them
-  codeMirrorAPI.forEach((funcName) => {
-    // Some functions that we want to proxy (like phrase) are not on the codemirror
-    // editor object when this code executes, so we have to do the lookup inside the
-    // wrapper function. Hopefully by the time the wrapper function is called,
-    // the function it proxies to has been added to the editor instance.
-    base[funcName] = (...args: unknown[]) =>
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      (editor.codemirror as any)[funcName](...args);
-  });
-
-  const api: ToggleEditorAPI = {
-    // custom CMB methods
-    getBlockMode: () => selectors.isBlockModeEnabled(store.getState()),
-    setBlockMode: (blockMode: boolean) => {
-      store.dispatch(actions.setBlockMode(blockMode, editor, language));
-    },
-    getCM: () => editor.codemirror,
-    on: (...args: Parameters<CodeMirror.Editor["on"]>) => {
-      const [type, fn] = args;
-      if (!eventHandlers[type]) {
-        eventHandlers[type] = [fn];
-      } else {
-        eventHandlers[type].push(fn);
-      }
-      editor.codemirror.on(type, fn);
-    },
-    off: (...args: Parameters<CodeMirror.Editor["on"]>) => {
-      const [type, fn] = args;
-      eventHandlers[type]?.filter((h) => h !== fn);
-      editor.codemirror.off(type, fn);
-    },
-    runMode: () => {
-      throw "runMode is not supported in CodeMirror-blocks";
-    },
-  };
-  return { ...base, ...api };
 };
 
 export type ToggleEditorProps = {
@@ -236,7 +75,7 @@ function ToggleEditor(props: ToggleEditorProps) {
    * API with mode-specific versions, (2) re-assign event handlers,
    * and (3) re-render any TextMarkers.
    */
-  const handleEditorMounted = (editor: CodeMirrorFacade, api: API) => {
+  const handleEditorMounted = (editor: CodeMirrorFacade) => {
     // set CM aria attributes, and mount announcer
     const mode = blockMode ? "Block" : "Text";
     const wrapper = editor.codemirror.getWrapperElement();
@@ -244,10 +83,9 @@ function ToggleEditor(props: ToggleEditorProps) {
     wrapper.setAttribute("aria-label", mode + " Editor");
     mountAnnouncer(wrapper);
     // Rebuild the API and assign re-events
-    props.onMount({
-      ...buildAPI(editor, store, props.language, eventHandlersRef.current),
-      ...api,
-    });
+    props.onMount(
+      buildAPI(editor, store, props.language, eventHandlersRef.current)
+    );
     for (const [type, handlers] of Object.entries(eventHandlersRef.current)) {
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
       handlers.forEach((h) => editor.codemirror.on(type as any, h));
